starwars_agent/pyproject.toml
[project]
name = "starwars-agent"
version = "0.1.0"
description = "Star Wars Fandom Knowledge Agent (Strands SDK + MongoDB vector search)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "a2a-sdk[all]",
  "strands-agents",
  "strands-agents-tools",
  "fastapi",
  "uvicorn",
  "httpx",
  "pymongo",
  "openai",
  "shared",
  "numpy>=2.4.2",
  "certifi",
]

[tool.uv]
package = true

[tool.uv.sources]
shared = { path = "../shared", editable = true }

[dependency-groups]
dev = [
  "ruff",
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pytest-httpx>=0.35",
  "testcontainers[mongodb]",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "CPY",
  "D417",
  "D418",
  "D419",
  "DOC",
  "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "S101",     # assert usage is standard in tests
  "S311",     # pseudo-random generators (fine for test data)
  "PLR2004",  # magic values in comparisons
  "SLF001",   # private member access (testing internals)
  "D",        # missing docstrings
  "ANN",      # missing type annotations
  "ARG",      # unused function arguments (fixtures)
  "B017",     # blind exception in pytest.raises
  "PT011",    # broad pytest.raises without match
  "INP001",   # implicit namespace package (test dirs)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
